fix: moving popover content to internal
