+++ packages/react/src/components/internal/PopoverContent/PopoverContent.styles.ts
@@ -1,4 +1,4 @@
-import { css } from '../../styles';
+import { css } from '../../../styles';
 
 export const useStyles = css({
   backgroundColor: '$background-surface',
